+++ formio/models/formio_version.py
@@ -58,9 +58,12 @@ class Version(models.Model):
         base_translations = self.env['formio.translation'].search(domain)
         vals_list = []
         for rec in self:
+            sequence = max(rec.translation_ids.mapped('sequence'))
             for trans in base_translations:
                 if not rec.translation_ids.filtered(lambda t: t.base_translation_id.id == trans.id):
+                    sequence += 1
                     vals = {
+                        'sequence': sequence,
                         'formio_version_id': rec.id,
                         'base_translation_id': trans.id,
                         'lang_id': trans.lang_id.id,